refactor: use modern routing
