+++ src/layout/index.tsx
@@ -3,7 +3,9 @@ import MainHeader from './header/Header'
 import SideBar from './sidebar/SideBar'
 import styled from 'styled-components'
 import LayoutContext from './LayoutContext'
-import RoutesList from '../routes'
+import { Outlet } from 'react-router-dom'
+import { Suspense } from 'react'
+import Preloader from 'src/shared/Preloader'
 
 const { Content } = Layout
 
@@ -19,7 +21,7 @@ const StyledBody = styled.div`
   min-height: 360px;
 `
 
-function MainLayout() {
+export function MainLayout() {
   return (
     <StyledLayout className="main">
       <LayoutContext>
@@ -28,7 +30,9 @@ function MainLayout() {
           <MainHeader />
           <StyledContent>
             <StyledBody>
-              <RoutesList />
+              <Suspense fallback={<Preloader />}>
+                <Outlet />
+              </Suspense>
             </StyledBody>
           </StyledContent>
         </Layout>
@@ -36,5 +40,3 @@ function MainLayout() {
     </StyledLayout>
   )
 }
-
-export default MainLayout
